Move union of thunk deps into dedicated method
